reorganised models folders and names
