+++ src/App.tsx
@@ -10,23 +10,23 @@ import ChartViewer from "./components/Charts";
 import GridFull from "./components/Grids";
 import TransactionsList from "./components/Transactions";
 
-// Assets
+// Models
 import {
   TransactionContext,
-  ConfigurationContext,
   Transaction,
-  iTransactionError,
   ITransaction,
-} from "./utils/types";
+  ITransactionError,
+} from "./models/transaction";
+import { ConfigurationContext } from "./models/configuration";
 
 // RENDER
 export default observer(function App() {
   // Definitions
-  const allTransactions = React.useContext(TransactionContext);
+  const transactionsStore = React.useContext(TransactionContext);
   const config = React.useContext(ConfigurationContext);
   const [isFetchingData, setIsFetchingData] = React.useState<boolean>(false);
   const [transactionErrorList, setTransactionErrorList] = React.useState<
-    Array<iTransactionError>
+    Array<ITransactionError>
   >([]);
 
   // Helpers
@@ -40,7 +40,7 @@ export default observer(function App() {
       .then((res) => {
         setIsFetchingData(false);
 
-        const errorArr: Array<iTransactionError> = [];
+        const errorArr: Array<ITransactionError> = [];
         const transactions: Array<[string, ITransaction]> = [];
         const indexMem = new Set();
 
@@ -91,7 +91,7 @@ export default observer(function App() {
               }
             },
           );
-        allTransactions.setTransactions(transactions);
+        transactionsStore.setTransactions(transactions);
         setTransactionErrorList(errorArr); // ??? trigs useless rerender of App; will be cleaned during backend implementation
       })
       .catch((err) => {
@@ -106,85 +106,87 @@ export default observer(function App() {
   };
 
   // Loading
-  React.useEffect(getGSheetData, [allTransactions, config]); // ??? Depedency is correct?
+  React.useEffect(getGSheetData, [transactionsStore, config]); // ??? Depedency is correct?
 
   return (
     <>
-      <TransactionContext.Provider value={allTransactions}>
-        <header className="grid grid-cols-2 p-4 bg-mred">
-          <h1 className="self-center text-3xl text-white">
-            <a href="/">Mony mony</a>
-          </h1>
-          <nav className="self-center text-right">
-            <button
-              className="p-2 text-white hover:text-mred-light"
-              onClick={getGSheetData}
-            >
-              <RefreshIcon
-                className={`inline h-6 w-6 ${
-                  isFetchingData === true ? "animate-spin-slow" : ""
-                }`}
-              />
-            </button>
-            <NavLink
-              className="nav-button"
-              activeClassName="bg-mred-light text-white"
-              to="/transactions"
-            >
-              Transactions
-            </NavLink>
-            <NavLink
-              className="nav-button"
-              activeClassName="bg-mred-light text-white"
-              to="/grid"
-            >
-              Grid
-            </NavLink>
-            <NavLink
-              className="nav-button"
-              activeClassName="bg-mred-light text-white"
-              to="/chart"
-            >
-              Chart
-            </NavLink>
-          </nav>
-        </header>
+      <TransactionContext.Provider value={transactionsStore}>
+        <ConfigurationContext.Provider value={config}>
+          <header className="grid grid-cols-2 p-4 bg-mred">
+            <h1 className="self-center text-3xl text-white">
+              <a href="/">Mony mony</a>
+            </h1>
+            <nav className="self-center text-right">
+              <button
+                className="p-2 text-white hover:text-mred-light"
+                onClick={getGSheetData}
+              >
+                <RefreshIcon
+                  className={`inline h-6 w-6 ${
+                    isFetchingData === true ? "animate-spin-slow" : ""
+                  }`}
+                />
+              </button>
+              <NavLink
+                className="nav-button"
+                activeClassName="bg-mred-light text-white"
+                to="/transactions"
+              >
+                Transactions
+              </NavLink>
+              <NavLink
+                className="nav-button"
+                activeClassName="bg-mred-light text-white"
+                to="/grid"
+              >
+                Grid
+              </NavLink>
+              <NavLink
+                className="nav-button"
+                activeClassName="bg-mred-light text-white"
+                to="/chart"
+              >
+                Chart
+              </NavLink>
+            </nav>
+          </header>
 
-        <main className="flex-auto p-8 text-sm text-gray-700">
-          <Switch>
-            <Route path="/transactions">
-              <h2 className="section-title">Transactions</h2>
-              <TransactionsList
-                transactionErrorList={transactionErrorList}
-              ></TransactionsList>
-            </Route>
-            <Route path="/grid">
-              <h2 className="section-title">Grid</h2>
-              <GridFull></GridFull>
-            </Route>
-            <Route path="/chart">
-              <h2 className="section-title">Chart</h2>
-              <ChartViewer></ChartViewer>
-            </Route>
-            <Route path="/">
-              <h2 className="section-title">Chart</h2>
-              <ChartViewer></ChartViewer>
-            </Route>
-          </Switch>
-        </main>
+          <main className="flex-auto p-8 text-sm text-gray-700">
+            <Switch>
+              <Route path="/transactions">
+                <h2 className="section-title">Transactions</h2>
+                <TransactionsList
+                  transactionErrorList={transactionErrorList}
+                ></TransactionsList>
+              </Route>
+              <Route path="/grid">
+                <h2 className="section-title">Grid</h2>
+                <GridFull></GridFull>
+              </Route>
+              <Route path="/chart">
+                <h2 className="section-title">Chart</h2>
+                <ChartViewer></ChartViewer>
+              </Route>
+              <Route path="/">
+                <h2 className="section-title">Chart</h2>
+                <ChartViewer></ChartViewer>
+              </Route>
+            </Switch>
+          </main>
 
-        <footer className="p-4 text-white bg-mred">
-          © 2021 by ivanprobst (
-          <a
-            className="underline"
-            target="_blank"
-            href="https://github.com/ivanprobst/monymony"
-            rel="noopener noreferrer"
-          >
-            check out this project on GitHub
-          </a>
-          )
-        </footer>
+          <footer className="p-4 text-white bg-mred">
+            © 2021 by ivanprobst (
+            <a
+              className="underline"
+              target="_blank"
+              href="https://github.com/ivanprobst/monymony"
+              rel="noopener noreferrer"
+            >
+              check out this project on GitHub
+            </a>
+            )
+          </footer>
+        </ConfigurationContext.Provider>
       </TransactionContext.Provider>
     </>
   );
